Fix to do dispatch agent if mapped scene is not active when launching autopilot
Situation:

Launch autopilot while in play mode, or player build.
And multiple scenes are loaded, but the scene the agent is mapped to is not active (it is loaded).

Problem:

Did not dispatch agent mapped to a loaded scene.

Fixes:

Check mapping all loaded scenes if active scene is not mapped on launch.